backend/requirements.txt
astroid==2.4.2
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.1
pylint==2.6.0
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
